# crates/crw-server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRW REST API",
    "version": "0.21.3",
    "description": "Firecrawl-compatible web scraping, crawling, mapping, search, and extraction API. Hosted at https://api.fastcrw.com; self-host with `docker run -p 3000:3000 ghcr.io/us/crw:latest`.",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://opensource.org/licenses/AGPL-3.0"
    },
    "contact": {
      "url": "https://github.com/us/crw"
    }
  },
  "servers": [
    {
      "url": "https://api.fastcrw.com",
      "description": "Hosted (requires Bearer token)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Self-hosted (no auth by default)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {}
  ],
...
  "paths": {
...
    "/v1/scrape": {
      "post": {
        "tags": [
          "scrape"
        ],
        "summary": "Scrape one URL",
        "operationId": "scrape",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scrape result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/v1/extract": {
      "post": {
        "tags": [
          "extract"
        ],
        "summary": "Start an async multi-URL structured extraction job",
        "description": "Native extraction. Returns a per-URL array via GET /v1/extract/{id}. At least one of `prompt` or `schema` is required.",
        "operationId": "extractStart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extract job accepted; poll with /v1/extract/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/extract/{id}": {
      "get": {
        "tags": [
          "extract"
        ],
        "summary": "Get extract job status and per-URL results",
        "operationId": "extractStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Extract job id returned by POST /v1/extract"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status + per-URL extraction results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractStatus"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Hosted only. Get a key at https://fastcrw.com/register."
      }
    },
    "schemas": {
...
      "ScrapeRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "formats": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "markdown",
                "html",
                "rawHtml",
                "plainText",
                "links",
                "json",
                "summary",
                "changeTracking"
              ]
            },
            "default": [
              "markdown"
            ]
          },
          "onlyMainContent": {
            "type": "boolean",
            "default": true,
            "description": "Strip navigation, ads, and boilerplate; keep main article content only"
          },
          "renderJs": {
            "type": "boolean",
            "nullable": true,
            "description": "null = auto-detect, true = force JS rendering, false = skip JS rendering"
          },
          "waitFor": {
            "type": "integer",
            "minimum": 0,
            "description": "Milliseconds to wait after JS rendering before extracting content"
          },
          "includeTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [],
            "description": "CSS selectors / HTML tags to include in extraction"
          },
          "excludeTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [],
            "description": "CSS selectors / HTML tags to remove before extraction"
          },
          "jsonSchema": {
            "type": "object",
            "additionalProperties": true,
            "description": "JSON Schema describing the structured output shape for the 'json' format"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Custom HTTP request headers forwarded to the target URL"
          },
          "cssSelector": {
            "type": "string",
            "description": "CSS selector to narrow content extraction to a specific element"
          },
          "xpath": {
            "type": "string",
            "description": "XPath expression to narrow content extraction to a specific element"
          },
          "chunkStrategy": {
            "$ref": "#/components/schemas/ChunkStrategy",
            "description": "Strategy for chunking the extracted markdown"
          },
          "query": {
            "type": "string",
            "description": "Query string for BM25/cosine chunk filtering (used with chunkStrategy)"
          },
          "filterMode": {
            "type": "string",
            "enum": [
              "bm25",
              "cosine"
            ],
            "description": "Filtering algorithm to rank chunks against query"
          },
          "topK": {
            "type": "integer",
            "minimum": 1,
            "description": "Number of top chunks to return after filtering (default: 5)"
          },
          "proxy": {
            "type": "string",
            "description": "Per-request proxy URL (http/https/socks5). Overrides global config."
          },
          "proxyList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [],
            "description": "Per-request proxy pool to rotate among (BYOP). Takes precedence over proxy and server config."
          },
          "proxyRotation": {
            "type": "string",
            "enum": [
              "round_robin",
              "random",
              "sticky_per_host"
            ],
            "description": "Rotation strategy for proxyList (default: sticky_per_host)"
          },
          "country": {
            "type": "string",
            "description": "2-letter ISO 3166-1 alpha-2 country code for residential proxy egress (e.g. 'us', 'gb')"
          },
          "stealth": {
            "type": "boolean",
            "description": "Override stealth mode for this request (null = use global config)"
          },
          "actions": {
            "type": "object",
            "additionalProperties": true,
            "description": "Unsupported Firecrawl parameter \u2014 returns a clear error if supplied"
          },
          "extract": {
            "$ref": "#/components/schemas/ExtractOptions",
            "description": "Firecrawl-compatible extract object (e.g. { schema: {...} })"
          },
          "llmApiKey": {
            "type": "string",
            "description": "BYOK: LLM provider API key for structured extraction (overrides server key)"
          },
          "llmProvider": {
            "type": "string",
            "enum": [
              "anthropic",
              "openai",
              "deepseek",
              "azure",
              "openai-compatible"
            ],
            "description": "Per-request LLM provider override"
          },
          "llmModel": {
            "type": "string",
            "description": "Per-request LLM model identifier override"
          },
          "baseUrl": {
            "type": "string",
            "description": "Per-request LLM base URL override for OpenAI-compatible providers (e.g. https://api.deepseek.com/v1)"
          },
          "summaryPrompt": {
            "type": "string",
            "maxLength": 500,
            "description": "User-supplied instructions appended to the summary system prompt (capped at 500 chars)"
          },
          "maxContentChars": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum bytes of scraped content sent to LLM for 'summary' format (capped at 200 KB)"
          },
          "renderer": {
            "type": "string",
            "enum": [
              "auto",
              "lightpanda",
              "chrome",
              "chrome_proxy",
              "playwright"
            ],
            "description": "Pin this request to a specific renderer. Non-auto implies renderJs:true unless renderJs:false is set."
          },
          "deadlineMs": {
            "type": "integer",
            "minimum": 1,
            "maximum": 60000,
            "description": "End-to-end deadline budget in milliseconds (default: 8000, max: 60000)"
          },
          "debug": {
            "type": "boolean",
            "description": "When true, response includes debugExtraction trace with every candidate the extractor considered"
          },
          "changeTracking": {
            "$ref": "#/components/schemas/ChangeTrackingOptions",
            "description": "Activated when formats includes 'changeTracking'. The format entry is the plain string; options ride on this sibling field."
          },
          "goal": {
            "type": "string",
            "maxLength": 2048,
            "description": "Plain-language monitor goal for the meaningful-change judge"
          },
          "judgeEnabled": {
            "type": "boolean",
            "description": "Run the LLM meaningful-change judge on a changed page (requires goal)"
          },
          "parsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParserSpec"
            },
            "description": "Document parser directives. Omit = auto-parse PDFs. Empty array = disable parsing."
          }
        }
      },
      "ScrapeResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "warning": {
            "type": "string",
            "description": "Top-level advisory (e.g. fallback render notice)"
          },
          "data": {
            "type": "object",
            "properties": {
              "markdown": {
                "type": "string"
              },
              "html": {
                "type": "string"
              },
              "rawHtml": {
                "type": "string"
              },
              "links": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "json": {
                "type": "object",
                "additionalProperties": true
              },
              "metadata": {
                "$ref": "#/components/schemas/PageMetadata"
              },
              "warning": {
                "type": "string"
              },
              "warnings": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "renderDecision": {
                "type": "string",
                "description": "Which renderer ran (e.g. 'http', 'browser')"
              },
              "creditCost": {
                "type": "number"
              }
            }
          }
        }
      },
...
      "ExtractRequest": {
        "type": "object",
        "required": [
          "urls"
        ],
        "description": "At least one of `prompt` or `schema` must be present.",
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "description": "URLs to extract from (1..crawler.max_extract_urls)."
          },
          "prompt": {
            "type": "string",
            "description": "Free-text extraction objective."
          },
          "schema": {
            "type": "object",
            "additionalProperties": true,
            "description": "JSON Schema constraining the extracted output."
          },
          "llmApiKey": {
            "type": "string",
            "description": "BYOK LLM API key."
          },
          "llmProvider": {
            "type": "string"
          },
          "llmModel": {
            "type": "string"
          },
          "basis": {
            "type": "boolean",
            "description": "Per-field attribution (Phase 2b). Rejected with 400 while unsupported."
          }
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "sourceURL": {
            "type": "string",
            "format": "uri"
          },
          "statusCode": {
            "type": "integer"
          }
        }
      },
      "ChangeTrackingOptions": {
        "type": "object",
        "properties": {
          "modes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "gitDiff",
                "json"
              ]
            },
            "description": "Diff surfaces. gitDiff = markdown unified diff + AST; json = per-field diff; both = mixed."
          },
          "schema": {
            "type": "object",
            "additionalProperties": true,
            "description": "JSON schema of tracked fields (json/mixed mode)"
          },
          "prompt": {
            "type": "string"
          },
          "previous": {
            "$ref": "#/components/schemas/ChangeTrackingSnapshot"
          },
          "tag": {
            "type": "string",
            "description": "Opaque caller tag echoed on the result"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type; non-text content is hashed, not diffed"
          }
        }
      },
      "ChangeTrackingSnapshot": {
        "type": "object",
        "properties": {
          "markdown": {
            "type": "string",
            "description": "Present for gitDiff/mixed mode"
          },
          "json": {
            "type": "object",
            "additionalProperties": true,
            "description": "Present for json/mixed mode"
          },
          "contentHash": {
            "type": "string",
            "description": "Mode-aware hash; persist + supply on the next check"
          },
          "capturedAt": {
            "type": "string",
            "description": "Caller-stamped capture time, echoed untouched"
          }
        }
      },
...
      "ChunkStrategy": {
        "type": "object",
        "description": "Strategy for chunking extracted markdown before query-based filtering",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "sentence",
              "regex",
              "topic"
            ],
            "description": "Chunking algorithm: sentence = split on .!?, regex = split on pattern, topic = split on markdown headings"
          },
          "maxChars": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum characters per chunk (merges short chunks up to this limit)"
          },
          "overlapChars": {
            "type": "integer",
            "minimum": 0,
            "description": "Character overlap between adjacent chunks"
          },
          "dedupe": {
            "type": "boolean",
            "description": "Remove duplicate chunks"
          },
          "pattern": {
            "type": "string",
            "description": "Regex pattern to split on (required when type = 'regex')"
          }
        }
      },
      "ExtractOptions": {
        "type": "object",
        "description": "Firecrawl-compatible extraction options sub-object",
        "properties": {
          "schema": {
            "type": "object",
            "additionalProperties": true,
            "description": "JSON Schema for structured extraction. Used as jsonSchema when jsonSchema is not set directly."
          }
        }
      },
      "ParserSpec": {
        "type": "object",
        "description": "Document parser directive. Accepts string form ('pdf') or object form.",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "pdf"
            ],
            "description": "Parser type (only 'pdf' is supported)"
          },
          "mode": {
            "type": "string",
            "enum": [
              "auto",
              "fast",
              "ocr"
            ],
            "description": "Parsing strategy (auto = text-first; ocr degrades to text extraction with a warning)"
          },
          "maxPages": {
            "type": "integer",
            "minimum": 1,
            "description": "Optional cap on the number of pages to parse"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "errorCode": {
            "type": "string",
            "description": "Machine-readable error code (see docs/error-codes.md)"
          }
        }
      },
      "ExtractAccepted": {
        "type": "object",
        "required": [
          "id",
          "status",
          "urls"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing"
            ]
          },
          "urls": {
            "type": "integer",
            "description": "Count of URLs enqueued for fetch (preflight-failed URLs appear in results, not here)."
          }
        }
      },
...
      "ExtractStatus": {
        "type": "object",
        "required": [
          "id",
          "status",
          "expiresAt",
          "creditsUsed",
          "tokensUsed"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing",
              "completed",
              "failed"
            ]
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtractUrlResult"
            },
            "description": "Per-URL results in original request order."
          },
          "error": {
            "type": "string",
            "description": "Job-level error, set only when every URL failed."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "creditsUsed": {
            "type": "integer"
          },
          "tokensUsed": {
            "type": "integer"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid bearer token (hosted only)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}